chore: Attempt to fix linter warnings.
Fixed several "assert" missing messages.
Unneeded "show" clauses
And a few hashCode calls which could be turned into a single line.

No functional changes.  Just using the linter and attempting to drive the
number of lints to zero (as part of learning the codebase).

This barely scratches the surface of the outstanding lints, but I didn't
want to make the PR too large to review.